refactor(cli): extract commands/search.ts and indexing.ts (batch 5)
- commands/search.ts: search/vectorSearch/querySearch plus
  parseStructuredQuery (lex:/vec:/hyde:/intent: query documents) and
  resolveCollectionFilter (also used by parseCLI's embed -c validation).
  sanitizeFTS5Term/buildFTS5Query/normalizeBM25 move along as-is —
  pre-existing dead code, kept per pure-move discipline.
- commands/indexing.ts: indexFiles/updateCollections/vectorIndex and
  the embed option parsers. Imports formatCount/shortModelName from
  doctor.ts (one-way edge).
- collectionAdd completes commands/collections.ts, now that indexFiles
  exists in indexing.ts (collections → indexing, one-way).
- Two misplaced comments dropped: a computeDisplayPath doc above
  updateCollections and a detectCollectionFromPath doc above
  collectionAdd — both described functions living elsewhere.

qmd.ts is now 999 lines: imports, finishSuccessfulCliCommand, parseCLI,
exitWithError, and the isMain dispatch switch.

Verified: tsc clean, build ok, 32 snapshots byte-identical,
cli/lazy-llm-import/exit-lifecycle green, cli.test.ts 146/146.
